requirements.txt
python-dateutil==2.8.2
Flask==2.0.1
psycopg2-binary==2.9.1